refactored the images in to another file
